app/add logging(+refactor) project service (#34)

+++ src/main/java/com/literandltx/taskmcapp/service/app/impl/ProjectServiceImpl.java
@@ -10,10 +10,13 @@ import com.literandltx.taskmcapp.repository.ProjectRepository;
 import com.literandltx.taskmcapp.service.app.ProjectService;
 import jakarta.persistence.EntityNotFoundException;
 import java.util.List;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+@Slf4j
 @RequiredArgsConstructor
 @Service
 public class ProjectServiceImpl implements ProjectService {
@@ -29,6 +32,10 @@ public class ProjectServiceImpl implements ProjectService {
         model.setUser(user);
         model.setStatus(Project.Status.INITIATED);
 
+        log.info(String.format(
+                "Project: %s, created for user: %s.",
+                requestDto.getName(), user.getUsername()));
+
         return projectMapper.toDto(projectRepository.save(model));
     }
 
@@ -37,9 +44,15 @@ public class ProjectServiceImpl implements ProjectService {
             final Pageable pageable,
             final User user
     ) {
-        return projectRepository.findAllByUser(pageable, user).stream()
+        List<ProjectRespondDto> list = projectRepository.findAllByUser(pageable, user).stream()
                 .map(projectMapper::toDto)
                 .toList();
+
+        log.info(String.format(
+                "User: %s, find all projects",
+                user.getUsername()));
+
+        return list;
     }
 
     @Override
@@ -47,9 +60,15 @@ public class ProjectServiceImpl implements ProjectService {
             final Long id,
             final User user
     ) {
-        return projectMapper.toDto(projectRepository.findByIdAndUserId(id, user.getId())
-                .orElseThrow(
-                        () -> new EntityNotFoundException("Cannot find project with id: " + id)));
+        Project project = projectRepository.findByIdAndUserId(id, user.getId()).orElseThrow(
+                () -> new EntityNotFoundException("Cannot find project with id: " + id));
+        ProjectRespondDto dto = projectMapper.toDto(project);
+
+        log.info(String.format(
+                "User: %s, find project by id: %s.",
+                user.getUsername(), id));
+
+        return dto;
     }
 
     @Override
@@ -59,6 +78,9 @@ public class ProjectServiceImpl implements ProjectService {
             final User user
     ) {
         if (!projectRepository.existsByIdAndUser(id, user)) {
+            log.info(String.format(
+                    "User: %s, tried to update project by id: %s, but failed. Cannot find project",
+                    user.getUsername(), id));
             throw new EntityNotFoundException("Cannot find project with id: " + id);
         }
 
@@ -67,6 +89,11 @@ public class ProjectServiceImpl implements ProjectService {
         model.setUser(user);
 
         final Project saved = projectRepository.save(model);
+
+        log.info(String.format(
+                "User: %s, updated project by id: %s, successfully",
+                user.getUsername(), id));
+
         return projectMapper.toDto(saved);
     }
 
@@ -75,6 +102,21 @@ public class ProjectServiceImpl implements ProjectService {
             final Long id,
             final User user
     ) {
+        Optional<Project> project = projectRepository.findByIdAndUserId(id, user.getId());
+
+        if (project.isEmpty()) {
+            log.info(String.format(
+                    "User: %s, tried to delete projectId: %s, but project do not exists.",
+                    user.getUsername(), id));
+            throw new RuntimeException(String.format(
+                    "Project with id: %s, do not exists.",
+                    id));
+        }
+
         projectRepository.deleteById(id);
+
+        log.info(String.format(
+                "User: %s, deleted projectId: %s, successfully",
+                user.getUsername(), id));
     }
 }
